% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMED_score.R
\name{aMED_score}
\alias{aMED_score}
\title{Comute the aMED score based on the foodgroup median intake. 
Below median intake is scored 0
Equal to or above intake is scored 1
For alcohol the score is based on a range
Female is scored 1 if between 5-25, otherwise 0
Male is scored 1 if between 10-50, otherwise 0}
\usage{
aMED_score(row, median_list)
}
\arguments{
\item{row}{A row from the df}

\item{median_list}{List of FoodGroups with their corresponding median values}
}
\value{
Score of 1 or 0 based on median foodgroup intake and individual intake
}
\description{
Comute the aMED score based on the foodgroup median intake. 
Below median intake is scored 0
Equal to or above intake is scored 1
For alcohol the score is based on a range
Female is scored 1 if between 5-25, otherwise 0
Male is scored 1 if between 10-50, otherwise 0
}
\author{
Robbie Pope
}
