% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_score.R
\name{animal_score}
\alias{animal_score}
\title{Compute animal score based on food group '+ve' or '-ve'. Positive or negative 
groups are assigned points based on quintiles as follows:
+ve Group: quintile 5 = 5, 
-ve Group: quintile 5 = 1}
\usage{
animal_score(df)
}
\arguments{
\item{df}{<- DF with FFQ_ID, and columns 'HealthfulGroup' and 'Quintile'}
}
\value{
score_df <- Df with additional column with quintile based animal scoring
}
\description{
Compute animal score based on food group '+ve' or '-ve'. Positive or negative 
groups are assigned points based on quintiles as follows:
+ve Group: quintile 5 = 5, 
-ve Group: quintile 5 = 1
}
\author{
Robbie Pope
}
