% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_animal_group.R
\name{define_animal_group}
\alias{define_animal_group}
\title{xx}
\usage{
define_animal_group(df)
}
\arguments{
\item{df}{<- DF with amalgamated FFQ groups in a column called 'FoodGroup'}
}
\value{
animal_df <- DF with additional column 'HealthfulGroup' with healthful group assignment
}
\description{
xx
}
\author{
Robbie Pope
}
