% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_food_group.R
\name{define_food_group}
\alias{define_food_group}
\title{Group the DataFrame by ID and calculate the sum of meal_portion for
specified meal_ID codes. Any FFQ IDs with no meal_id data are added as 0g.}
\usage{
define_food_group(df, id_list, food_group_name)
}
\arguments{
\item{df}{<- FFQ Line FETA processed df of FFQ_IDs, MEAL_IDs and MEAL_Portions}

\item{id_list}{<- Vector of numerical meal_ids for a food group}

\item{food_group_name}{<- String, name of food group, e.g., "animal_fat"}
}
\value{
summarized df for food group
}
\description{
Group the DataFrame by ID and calculate the sum of meal_portion for
specified meal_ID codes. Any FFQ IDs with no meal_id data are added as 0g.
}
\author{
Robbie Pope
}
