% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.adj.residual.R
\name{energy.adj.residual}
\alias{energy.adj.residual}
\title{Function to compute nutrient residual energy-adjusted model and return df
with adjusted columns.}
\usage{
energy.adj.residual(df, energy_col, start_col, end_col)
}
\arguments{
\item{df}{Dataframe of FFQ processed data containing energy in kcal, all NA should be 0}

\item{energy_col}{The index of the energy_kcal column to use as a mean}

\item{start_col}{The index of the first nutrient/food group column}

\item{end_col}{The index of the final nutrient/food group column}
}
\value{
Dataframe with nutrient/food group energy adjusted additional columns
}
\description{
Energy adjusted intake = residual + prediction from mean kcal
}
\details{
Methodology: Adjustmentfor total energy intake in epidemiologic studies,  Willet et al. 1997
}
\author{
Robbie Pope
}
